Update Kafka Test case
